+++ lib/ooxml_parser/docx_parser/docx_data/document_structure/docx_paragraph/docx_formula/pre_sub_superscript.rb
@@ -1,0 +1,27 @@
+module OoxmlParser
+  # Class for parsing `m:sPre` object
+  class PreSubSuperscript < OOXMLDocumentObject
+    # @return [DocxFormula] top value
+    attr_accessor :top_value
+
+    # @return [DocxFormula] bottom_value
+    attr_accessor :bottom_value
+
+    # @return [DocxFormula] base
+    attr_accessor :base
+
+    def parse(node)
+      node.xpath('*').each do |node_child|
+        case node_child.name
+        when 'sub'
+          @bottom_value = DocxFormula.parse(node_child)
+        when 'sup'
+          @top_value = DocxFormula.parse(node_child)
+        when 'e'
+          @base = DocxFormula.parse(node_child)
+        end
+      end
+      self
+    end
+  end
+end